+++ browser-extension/src/components/instructions/ModelDownloadButton.vue
