+++ browser-extension/src/entrypoints/instructions/index.html
@@ -6,7 +6,7 @@
       name="viewport"
       content="width=device-width, initial-scale=1.0"
     />
-    <title>AIGELB - Information</title>
+    <title>AIGELB - Anleitung</title>
   </head>
   <body>
     <div id="app"></div>
