+++ browser-extension/src/entrypoints/popup/App.vue
@@ -1,23 +1,9 @@
 <template>
   <v-app style="width: 800px">
-    <v-container>
-      <v-btn
-        :prepend-icon="mdiAccessPoint"
-        @click="sendMessageTest"
-        >Message test</v-btn
-      >
-    </v-container>
+    <v-container> </v-container>
   </v-app>
 </template>
 
 <script setup lang="ts">
-import { mdiAccessPoint } from "@mdi/js";
-import { VApp, VBtn, VContainer } from "vuetify/components";
-
-import { sendMessage } from "@/utility/messaging.ts";
-
-async function sendMessageTest() {
-  const length = await sendMessage("testEvent", "hello world");
-  console.debug(`POPUP - Received: ${length}`);
-}
+import { VApp, VContainer } from "vuetify/components";
 </script>
